Move front end updater code to own file, update electron error message
