+++ src/index.js
@@ -1,5 +1,3 @@
-/* eslint-disable no-console */
-
 import React from 'react';
 import ReactDOM from 'react-dom';
 import injectTapEventPlugin from 'react-tap-event-plugin';
@@ -9,26 +7,10 @@ import { Provider } from 'react-redux';
 import { store } from './ducks/store';
 import App from './containers/App';
 import AppRouter from './routes';
-import { isCordova, isElectron } from './utils/Environment';
+import { isCordova } from './utils/Environment';
+import updater from './utils/updater';
 
 injectTapEventPlugin();
-
-// Initialise auto update if we are in electron
-const checkForUpdate = () => {
-  if (isElectron()) {
-    // update through IPC goes here
-    const { ipcRenderer } = window.require('electron');
-    ipcRenderer.send('CHECK_FOR_UPDATE');
-
-    ipcRenderer.on('UPDATE_FOUND', (event, arg) => {
-      console.log(arg);
-    });
-
-    ipcRenderer.on('UP_TO_DATE', (event, arg) => {
-      console.log(arg);
-    });
-  }
-};
 
 const startApp = () => {
   ReactDOM.render(
@@ -40,7 +22,7 @@ const startApp = () => {
     document.getElementById('root'),
   );
 
-  checkForUpdate();
+  updater.checkForUpdate();
 };
 
 if (isCordova()) {
